#18 Move Jira services to service.jira package, extract LoginResponseParser
