+++ src/main/java/com/senla/worklog/reminder/service/jira/DefaultLoginResponseParser.java
@@ -1,0 +1,62 @@
+package com.senla.worklog.reminder.service.jira;
+
+import com.senla.worklog.reminder.exception.JiraAuthenticationException;
+import com.senla.worklog.reminder.model.JiraSession;
+import org.jsoup.Jsoup;
+import org.jsoup.nodes.Document;
+import org.jsoup.nodes.Element;
+import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Component;
+
+import java.util.List;
+
+import static liquibase.repackaged.org.apache.commons.lang3.StringUtils.substringBefore;
+
+@Component
+public class DefaultLoginResponseParser implements LoginResponseParser {
+    private static final String UNEXPECTED_ERROR_MESSAGE = "Unexpected error occurred";
+
+    @Override
+    public JiraSession parseLoginResponse(ResponseEntity<String> response) {
+        validateLoginResponse(response);
+        return getJiraSession(response);
+    }
+
+    private JiraSession getJiraSession(ResponseEntity<String> response) {
+        List<String> cookies = getCookies(response);
+        String sessionId = findSessionId(cookies);
+        return new JiraSession(sessionId);
+    }
+
+    private List<String> getCookies(ResponseEntity<String> response) {
+        List<String> setCookies = response.getHeaders().get("Set-Cookie");
+        if (setCookies == null) {
+            throw new JiraAuthenticationException("Response on login request does not contain Set-Cookie header");
+        }
+        return setCookies;
+    }
+
+    private String findSessionId(List<String> cookies) {
+        return cookies.stream()
+                .filter(cookie -> cookie.startsWith("JSESSIONID"))
+                .map(cookie -> substringBefore(cookie, ";"))
+                .findFirst()
+                .orElseThrow(() -> new JiraAuthenticationException("JSESSIONID not found in Set-Cookie login response"));
+    }
+
+    private void validateLoginResponse(ResponseEntity<String> response) {
+        if (!response.getStatusCode().is3xxRedirection()) {
+            String errorMessage = resolveLoginErrorMessage(response.getBody());
+            throw new JiraAuthenticationException("Login failed; error message: " + errorMessage);
+        }
+    }
+
+    private String resolveLoginErrorMessage(String responseBody) {
+        Document document = Jsoup.parse(responseBody);
+        Element formBody = document.getElementsByClass("form-body").first();
+        if (formBody == null) {
+            return UNEXPECTED_ERROR_MESSAGE;
+        }
+        return formBody.getElementsByClass("aui-message error").text();
+    }
+}
