+++ src/main/java/com/senla/worklog/reminder/service/jira/JiraAuthenticationService.java
@@ -1,4 +1,4 @@
-package com.senla.worklog.reminder.service;
+package com.senla.worklog.reminder.service.jira;
 
 import com.senla.worklog.reminder.model.JiraSession;
 import org.springframework.http.HttpHeaders;
