+++ src/main/java/com/senla/worklog/reminder/service/jira/LoginResponseParser.java
@@ -1,0 +1,8 @@
+package com.senla.worklog.reminder.service.jira;
+
+import com.senla.worklog.reminder.model.JiraSession;
+import org.springframework.http.ResponseEntity;
+
+public interface LoginResponseParser {
+    JiraSession parseLoginResponse(ResponseEntity<String> response);
+}
